Add GLUELESS to support glueless OMR build.
By default OMR requires a 'glue' to describe the language interface
for things like garbage collection. But that doesn't make that much
sense in a world where we are building without the GC.

So this enables GLUELESS functionality. This has not yet been
merged onto the official OMR however, so requires my fork at the moment.