Part 1: Prepare to eliminate genCall_ruby_stack
The genCall_ruby_stack function funnels a number distinct VM linkages
through a common system. This is mostly fine for 2.2, as the linkages
share a number of commonalities. However, in Ruby 2.3, especially after
d5ec9ec, this commonality starts go away, and we begin to be excessively
pressured by the funnel.

This is part I of multiple commits on this road.